move ContactInfoLinkFilter to a new file
