+++ extract_emails/link_filters/__init__.py
@@ -1,2 +1,3 @@
 from .link_filter_interface import LinkFilterInterface
-from .default_link_filter import DefaultLinkFilter, ContactInfoLinkFilter
+from .default_link_filter import DefaultLinkFilter
+from .contact_link_filter import ContactInfoLinkFilter
